Compile fix for OS X - don't try to print a size_t in assert macro

+++ renderdoc/driver/gl/wrappers/gl_emulated.cpp
@@ -1174,7 +1174,7 @@ void APIENTRY _glClearBufferSubData(GLenum target, GLenum internalformat, GLintp
 
     size_t stride = compCount * compByteWidth;
 
-    RDCASSERT(size % stride == 0, size, stride);
+    RDCASSERT(size % stride == 0, uint64_t(size), uint64_t(stride));
 
     // copy without conversion
     for(GLsizeiptr i = 0; i < size; i += stride)
